+++ src/main/java/kitchenpos/product/dto/ProductResponse.java
@@ -2,12 +2,10 @@ package kitchenpos.product.dto;
 
 import kitchenpos.product.domain.Product;
 
-import java.math.BigDecimal;
-
 public class ProductResponse {
     private Long id;
     private String name;
-    private BigDecimal price;
+    private long price;
 
     public ProductResponse() {
 
@@ -27,7 +25,7 @@ public class ProductResponse {
         return name;
     }
 
-    public BigDecimal getPrice() {
+    public long getPrice() {
         return price;
     }
 }
